PW-23(task):Fix codeclimate maitainability issues
