+++ .circleci/config.yml
@@ -11,17 +11,17 @@ jobs:
       # Download and cache dependencies
       - restore_cache:
           keys:
-            - v1-dependencies-{{ .Branch }}-{{ checksum "package.json" }}
+            - v1-dependencies-{{ .Branch }}-{{ checksum "yarn.lock" }}
             # fallback to using the latest cache if no exact match is found
             - v1-dependencies-
       - run:
           name: Install dependencies
-          command: yarn
+          command: yarn install 
 
       - save_cache:
           paths:
             - node_modules
-          key: v1-dependencies-{{ .Branch }}-{{ checksum "package.json" }}
+          key: v1-dependencies-{{ .Branch }}-{{ checksum "yarn.lock" }}
   
       # Download code climate reporter.
       - run:
