+++ .codeclimate.yml
@@ -1,0 +1,32 @@
+version: "2"
+
+checks:
+ file-lines:
+    enabled: true
+    config:
+      threshold: 250
+  similar-code:
+    enabled: false
+  method-lines:
+    enabled: true
+    config:
+      threshold: 250
+  similar-code:
+    enabled: true
+    config:
+      threshold: 5
+plugins:
+  eslint:
+    enabled: true
+
+exclude_patterns:
+  - "node_modules/"
+  - ".nyc_output"
+  - "coverage"
+  - "tailwind.*"
+  - "tailwindcss.*"
+  - "./src/index.js"
+  - "./src/*/__test__/*"
+  - "./src/*/*.test.*"
+  - "./src/api/api.js"
+  - ".circleci"
