--- codeclimate.yml
@@ -1,19 +1,0 @@
-engines:
-  duplication:
-    enabled: false
-    config:
-      languages: -javascript
-  eslint:
-    enabled: true
-    checks:
-      similar-code:
-        enabled: false
-exclude_patterns:
-  - "node_modules/"
-  - ".nyc_output"
-  - "coverage"
-  - "tailwind.*"
-  - "tailwindcss.*"
-  - "./src/index.js"
-  - "./src/*/test.js"
-  - "./src/api/api.js"
