+++ src/components/Portfolio.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import PropTypes from 'prop-types';
-import ProjectOverview from '../pages/ProjectsOverview';
-import ProjectCategoryPage from '../pages/ProjectCategoryPage';
+import ProjectOverview from '../pages/ProjectsPage';
+// import ProjectCategoryPage from '../pages/ProjectCategoryPage';
 
 function Portfolio({ match }) {
   return (
@@ -13,7 +13,7 @@ function Portfolio({ match }) {
         <Route
           exact
           path={`${match.path}/:categoryId`}
-          component={ProjectCategoryPage}
+          component={ProjectOverview}
         />
       </Switch>
     </Router>
