+++ src/pages/ProjectsPage.jsx
@@ -6,8 +6,10 @@ import Spinner from '../components/Spinner';
 import NotFound from './404';
 import Divider from '../components/Divider';
 import PropTypes from 'prop-types';
+import ProjectCard from './ProjectCard';
 
 function ProjectOverview({ match }) {
+  const id = match.params.categoryId;
   const dispatch = useDispatch();
   useEffect(() => {
     dispatch(getProjects());
@@ -26,18 +28,8 @@ function ProjectOverview({ match }) {
       <article>
         <Divider />
       </article>
-      <article className="flex flex-wrap px-8 items-center content-start">
-        {projects.map((project, i) => (
-          <div
-            key={i}
-            className="flex-1 flex-wrap h-36 p-2 w-60 mb-2 border rounded hover:shadow-md sm:mt-2 sm:w-60 sm:h-56 sm:p-8 mr-4 md:w-64 lg:w-80 xl:w-96 "
-          >
-            <div className="">{project.name}</div>
-            <div className="w-60 md:w-60 h-14 overflow-hidden">
-              {project.description}
-            </div>
-          </div>
-        ))}
+      <article>
+        <ProjectCard projects={projects} id={id} />
       </article>
     </section>
   );
